missed files

+++ backend/src/main/java/com/iitgn/entryexit/repositories/PrivilegeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface PrivilegeRepository extends JpaRepository<Privilege, Long> {
+public interface PrivilegeRepository extends JpaRepository<Privilege, Integer> {
     Optional<Privilege> findByName(String name);
 }
